Cleanup and support for major/minor revision patch line purging
